Fix issue where we require a different file than we check for
Prior to this patch this code checked for `autoload.php` but then tried to require `vendor/autoload.php`. I changed the string passed to `require` to match the string passed to `file_exists`.

Without this change I would get a fatal error as phack would `require` a file that didn't exist.